Fixed Department From Float to double

+++ src/main/kotlin/com/prasetia/erp/pojo/department/DepartmentSummary.kt
@@ -3,9 +3,9 @@ package com.prasetia.erp.pojo.department
 class DepartmentSummary(
         val id:Long?,
         val periode:Long?,
-        val nilai_budget:Long?,
-        val realisasi_budget:Long?,
-        val persent_budget:Float?
+        val nilai_budget:Double?,
+        val realisasi_budget:Double?,
+        val persent_budget:Double?
 ){
-    constructor(): this(0,0,0,0,0f)
+    constructor(): this(0,0,0.0,0.0,0.0)
 }
